EditableHTML inherits from Editable
