+++ app/components/Podcast.js
@@ -2,8 +2,7 @@ import React from 'react'
 import Animate from 'react-animate'
 import Sidebar from 'Sidebar.js'
 import {I18n} from 'I18n.js'
-import {Editable} from 'Editable.js'
-import {EditableHTML} from 'EditableHTML.js'
+import {Editable, EditableHTML} from 'Editable.js'
 
 let PodcastPlayer = Animate.extend(class PodcastPlayer extends React.Component {
   constructor() {
